basic_nsfw_detection/requirements.txt
flask
transformers
torch
pillow
